elimino y modifico productos segun rol admin/premium

+++ dao/dbManager/productManagerDB.js
@@ -6,13 +6,22 @@ class ProductsManagerDB {
 
   //CREO UN PRODUCTO 
   static async create(req, res) {
-    const { body } = req
-   /* const imagenProducto = {
+    const { body } = req;
+  
+    // Verificar si el usuario es premium o no
+    const isPremiumUser = req.user.role === 'premium';
+  
+    // Establecer el owner del producto según el tipo de usuario
+    const owner = isPremiumUser ? req.user.email : 'admin';
+  
+    // Agregar el campo owner al cuerpo del producto
+    const productData = {
       ...body,
-      //avatar: `/static/imgs/${file.originalname}`,
-    }*/
-    const result = await Products.createProduct(body)
-    res.status(201).json(result)
+      owner: owner
+    };
+  
+    const result = await Products.createProduct(productData);
+    res.status(201).json(result);
   }
   //LLAMO A LOS PRODUCTOS
   static async get(req, res) {
@@ -31,17 +40,47 @@ class ProductsManagerDB {
   }
   //MODIFICO UN PRODUCTO POR ID
   static async updateById(req, res) {
-   const id = req.params.id
-   const data = req.body
-    //const { params: { id }, data } = req
-    await Products.updateProductById(id, data)
-    res.status(240).json({message: "ok"})
+    const id = req.params.id;
+    const data = req.body;
+  
+    // Verificar si el usuario es premium o admin
+    const isAdmin = req.user.role === 'admin';
+    const isPremiumUser = req.user.role === 'premium';
+  
+    // Obtener el owner actual del producto
+    const product = await Products.getProductById(id);
+    if (!product) {
+      return res.status(404).end();
+    }
+  
+    // Verificar si el usuario puede actualizar el producto
+    if (!(isAdmin || (isPremiumUser && product.owner === req.user.email))) {
+      return res.status(403).json({ message: 'No tienes permiso para actualizar este producto' });
+    }
+  
+    await Products.updateProductById(id, data);
+    res.status(200).json({ message: 'Producto actualizado exitosamente' });
   }
   //ELIMINO UN PRODUCTO POR ID
   static async deleteById(req, res) {
-    const { params: { id } } = req
-    await Products.deleteProductById({ _id: id })
-    res.status(204).end()
+    const id = req.params.id;
+  
+    // Verificar si el usuario es admin
+    const isAdmin = req.user.role === 'admin';
+  
+    // Obtener el owner actual del producto
+    const product = await Products.getProductById(id);
+    if (!product) {
+      return res.status(404).end();
+    }
+  
+    // Verificar si el usuario puede eliminar el producto
+    if (!(isAdmin || product.owner === req.user.email)) {
+      return res.status(403).json({ message: 'No tienes permiso para eliminar este producto' });
+    }
+  
+    await Products.deleteProductById({ _id: id });
+    res.status(204).end();
   }
   //FILTRO POR CATEGORIA
   static async filtroCategory(req, res) {
